Fixed issues in the Serial class which caused data which had already been recieved to be re-recieved in part and, in many cases, miss the delimiters ending a string sent from a device.
The primary issue stemmed from a misunderstanding of the serialport callback
data. The returned value is an array of numbers, which represent the
characters recieved. All that is required is to use .toString() to convert
the values to a string. The array also included a subobject called .buffer
which is the internal buffer of data received, and not only the newest data.
We were taking the data out of the .buffer subobject and not just using the
array data and length.

Also, the use of split to parse out delimited strings allowed for the
possibility of the delimeter being missed if it came at the end of the string.
e.g. if you split "123\n45" you know the \n was there because you have an
array of length 2. But if you split "123\n" you get an array of length 1, and
if you split "123" you also get an array of length 1, with no indication that
the delimeter was present.